chore(sdk): Split the `list.rs` module into `list/mod.rs` and `list/builder.rs`.
